site/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "garrettleber"
version = "0.1.0"
description = "Flask + WebTUI version of garrettleber.com"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0",
    "pyyaml>=6.0",
    "markdown>=3.5",
    "pygments>=2.17",
    "frozen-flask>=1.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[project.scripts]
site = "app:create_app"
